requirements.txt
# ---------------------------------------------
# 📊 DATA PROCESSING & ANALYSIS
# ---------------------------------------------
pandas>=2.0.0          # Data manipulation and analysis — core for working with tabular data
numpy>=1.25.0          # Numerical operations, arrays, and matrix support (Python 3.12対応)

# ---------------------------------------------
# 🧠 MACHINE LEARNING
# ---------------------------------------------
scikit-learn>=1.4.0    # Classic ML models (regression, classification, clustering), preprocessing, model evaluation
xgboost>=1.7.5
# ---------------------------------------------
# 📈 VISUALIZATION
# ---------------------------------------------
matplotlib==3.7.1      # Base library for plotting graphs, charts, and visualizations
seaborn==0.12.2        # High-level interface for drawing attractive and informative statistical graphics (built on matplotlib)

# ---------------------------------------------
# 📦 EXPERIMENT TRACKING & MODEL MANAGEMENT
# ---------------------------------------------
mlflow>=2.10.0

# ---------------------------------------------
# ✅ TESTING
# ---------------------------------------------
pytest==7.3.1          # Python testing framework — great for writing unit tests for ML pipelines, data validation, etc.
pytest-cov==4.1.0      # Coverage reporting for pytest

# ---------------------------------------------
# ⚡ API DEVELOPMENT (FOR MODEL SERVING)
# ---------------------------------------------
fastapi==0.95.2        # Lightweight, high-performance web framework for serving ML models via REST APIs
uvicorn==0.22.0        # ASGI server for running FastAPI apps — lightweight and fast
#
# ---------------------------------------------
#  MISC
# ---------------------------------------------
pyyaml>=6.0.1
joblib==1.3.1
setuptools==65.5.0
ipykernel==6.29.5
pyarrow>=14.0.0

# ---------------------------------------------
# 🔧 CODE QUALITY & FORMATTING (CI/CD用)
# ---------------------------------------------
flake8==6.0.0          # Code linting and style checking
black==23.7.0          # Code formatting
isort==5.12.0          # Import sorting
mypy==1.5.1            # Static type checking
bandit==1.7.5          # Security linting
types-PyYAML>=6.0.12   # Type stubs for PyYAML (mypy用)


